requirements.txt
# Python dependencies for the embedding server with FAISS
flask>=2.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
torch>=2.0.0
faiss-cpu>=1.7.4
